Fixed decimal 2 precision

+++ src/main/java/explorecali/dto/response/RagingAverateDto.java
@@ -8,5 +8,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class RagingAverateDto {
-    private Long average;
+    private Double average;
 }
